move ui stuff to SubtitleEdit project.
