+++ libse/LibSE.csproj
@@ -513,7 +513,6 @@
     <Compile Include="UknownFormatImporterJson.cs" />
     <Compile Include="UknownFormatImporter.cs" />
     <Compile Include="Utilities.cs" />
-    <Compile Include="VideoInfo.cs" />
     <Compile Include="VobSub\Helper.cs" />
     <Compile Include="VobSub\Idx.cs" />
     <Compile Include="VobSub\IdxParagraph.cs" />
