+++ libse/Utilities.cs
@@ -25,86 +25,6 @@ namespace Nikse.SubtitleEdit.Core
         /// </summary>
         public static readonly char[] NewLineChars = { '\r', '\n' };
 
-        public static VideoInfo TryReadVideoInfoViaMatroskaHeader(string fileName)
-        {
-            var info = new VideoInfo { Success = false };
-            using (var matroska = new MatroskaFile(fileName))
-            {
-                if (matroska.IsValid)
-                {
-                    double frameRate;
-                    int width;
-                    int height;
-                    double milliseconds;
-                    string videoCodec;
-                    matroska.GetInfo(out frameRate, out width, out height, out milliseconds, out videoCodec);
-
-                    info.Width = width;
-                    info.Height = height;
-                    info.FramesPerSecond = frameRate;
-                    info.Success = true;
-                    info.TotalMilliseconds = milliseconds;
-                    info.TotalSeconds = milliseconds / TimeCode.BaseUnit;
-                    info.TotalFrames = info.TotalSeconds * frameRate;
-                    info.VideoCodec = videoCodec;
-                }
-            }
-            return info;
-        }
-
-        public static VideoInfo TryReadVideoInfoViaAviHeader(string fileName)
-        {
-            var info = new VideoInfo { Success = false };
-
-            try
-            {
-                using (var rp = new RiffParser())
-                {
-                    if (rp.TryOpenFile(fileName) && rp.FileType == RiffParser.ckidAVI)
-                    {
-                        var dh = new RiffDecodeHeader(rp);
-                        dh.ProcessMainAVI();
-                        info.FileType = RiffParser.FromFourCC(rp.FileType);
-                        info.Width = dh.Width;
-                        info.Height = dh.Height;
-                        info.FramesPerSecond = dh.FrameRate;
-                        info.TotalFrames = dh.TotalFrames;
-                        info.TotalMilliseconds = dh.TotalMilliseconds;
-                        info.TotalSeconds = info.TotalMilliseconds / TimeCode.BaseUnit;
-                        info.VideoCodec = dh.VideoHandler;
-                        info.Success = true;
-                    }
-                }
-            }
-            catch
-            {
-            }
-            return info;
-        }
-
-        public static VideoInfo TryReadVideoInfoViaMp4(string fileName)
-        {
-            var info = new VideoInfo { Success = false };
-
-            try
-            {
-                var mp4Parser = new MP4Parser(fileName);
-                if (mp4Parser.Moov != null && mp4Parser.VideoResolution.X > 0)
-                {
-                    info.Width = mp4Parser.VideoResolution.X;
-                    info.Height = mp4Parser.VideoResolution.Y;
-                    info.TotalMilliseconds = mp4Parser.Duration.TotalSeconds;
-                    info.VideoCodec = "MP4";
-                    info.FramesPerSecond = mp4Parser.FrameRate;
-                    info.Success = true;
-                }
-            }
-            catch
-            {
-            }
-            return info;
-        }
-
         // TODO: Change to IReadonlyList in .net >= 4.5
         public static ICollection<string> VideoFileExtensions { get; } = new List<string>
         { ".avi", ".mkv", ".wmv", ".mpg", ".mpeg", ".divx", ".mp4", ".asf", ".flv",".mov", ".m4v", ".vob", ".ogv", ".webm", ".ts", ".m2ts", ".avs", ".mxf" };
