+++ src/Logic/UiUtil.cs
@@ -1,5 +1,8 @@
 ﻿using Nikse.SubtitleEdit.Controls;
 using Nikse.SubtitleEdit.Core;
+using Nikse.SubtitleEdit.Core.ContainerFormats;
+using Nikse.SubtitleEdit.Core.ContainerFormats.Matroska;
+using Nikse.SubtitleEdit.Core.ContainerFormats.Mp4;
 using Nikse.SubtitleEdit.Core.SubtitleFormats;
 using Nikse.SubtitleEdit.Forms;
 using Nikse.SubtitleEdit.Logic.VideoPlayers;
@@ -20,15 +23,15 @@ namespace Nikse.SubtitleEdit.Logic
 
         public static VideoInfo GetVideoInfo(string fileName)
         {
-            VideoInfo info = Utilities.TryReadVideoInfoViaAviHeader(fileName);
+            VideoInfo info = TryReadVideoInfoViaAviHeader(fileName);
             if (info.Success)
                 return info;
 
-            info = Utilities.TryReadVideoInfoViaMatroskaHeader(fileName);
+            info = TryReadVideoInfoViaMatroskaHeader(fileName);
             if (info.Success)
                 return info;
 
-            info = Utilities.TryReadVideoInfoViaMp4(fileName);
+            info = TryReadVideoInfoViaMp4(fileName);
             if (info.Success)
                 return info;
 
@@ -42,7 +45,7 @@ namespace Nikse.SubtitleEdit.Logic
         private static VideoInfo TryReadVideoInfoViaDirectShow(string fileName)
         {
             return QuartsPlayer.GetVideoInfo(fileName);
-        }      
+        }
 
         public static int ShowSubtitle(Subtitle subtitle, VideoPlayerContainer videoPlayerContainer)
         {
@@ -651,5 +654,84 @@ namespace Nikse.SubtitleEdit.Logic
             return sb.ToString();
         }
 
+        public static VideoInfo TryReadVideoInfoViaMatroskaHeader(string fileName)
+        {
+            var info = new VideoInfo { Success = false };
+            using (var matroska = new MatroskaFile(fileName))
+            {
+                if (matroska.IsValid)
+                {
+                    double frameRate;
+                    int width;
+                    int height;
+                    double milliseconds;
+                    string videoCodec;
+                    matroska.GetInfo(out frameRate, out width, out height, out milliseconds, out videoCodec);
+
+                    info.Width = width;
+                    info.Height = height;
+                    info.FramesPerSecond = frameRate;
+                    info.Success = true;
+                    info.TotalMilliseconds = milliseconds;
+                    info.TotalSeconds = milliseconds / TimeCode.BaseUnit;
+                    info.TotalFrames = info.TotalSeconds * frameRate;
+                    info.VideoCodec = videoCodec;
+                }
+            }
+            return info;
+        }
+
+        public static VideoInfo TryReadVideoInfoViaAviHeader(string fileName)
+        {
+            var info = new VideoInfo { Success = false };
+
+            try
+            {
+                using (var rp = new RiffParser())
+                {
+                    if (rp.TryOpenFile(fileName) && rp.FileType == RiffParser.ckidAVI)
+                    {
+                        var dh = new RiffDecodeHeader(rp);
+                        dh.ProcessMainAVI();
+                        info.FileType = RiffParser.FromFourCC(rp.FileType);
+                        info.Width = dh.Width;
+                        info.Height = dh.Height;
+                        info.FramesPerSecond = dh.FrameRate;
+                        info.TotalFrames = dh.TotalFrames;
+                        info.TotalMilliseconds = dh.TotalMilliseconds;
+                        info.TotalSeconds = info.TotalMilliseconds / TimeCode.BaseUnit;
+                        info.VideoCodec = dh.VideoHandler;
+                        info.Success = true;
+                    }
+                }
+            }
+            catch
+            {
+            }
+            return info;
+        }
+
+        public static VideoInfo TryReadVideoInfoViaMp4(string fileName)
+        {
+            var info = new VideoInfo { Success = false };
+
+            try
+            {
+                var mp4Parser = new MP4Parser(fileName);
+                if (mp4Parser.Moov != null && mp4Parser.VideoResolution.X > 0)
+                {
+                    info.Width = mp4Parser.VideoResolution.X;
+                    info.Height = mp4Parser.VideoResolution.Y;
+                    info.TotalMilliseconds = mp4Parser.Duration.TotalSeconds;
+                    info.VideoCodec = "MP4";
+                    info.FramesPerSecond = mp4Parser.FrameRate;
+                    info.Success = true;
+                }
+            }
+            catch
+            {
+            }
+            return info;
+        }
     }
 }
